backend/requirements.txt
Django==3.0.5
django-cors-headers==3.2.0
django-ordered-model==3.3.0
djangorestframework==3.10.3
gunicorn==20.0.4
Markdown==3.1.1
Pillow==8.4.0
psycopg2-binary==2.8.4
pytz==2019.3
PyYAML==5.3.1
sentry-sdk==0.16.1
python-dotenv==0.12.0
huobi==0.1.12

pep8==1.7.1
autopep8==1.4.4
pycodestyle==2.5.0
requests==2.23.0
XlsxWriter==1.3.7
